+++ test/unit/fuzz-tools.test.js
@@ -503,6 +503,7 @@ import * as grayWealeV524 from '../../lib/gray-weale-v524.js';
 import * as nsofaV526 from '../../lib/nsofa-v526.js';
 import * as wayneIndexV527 from '../../lib/wayne-index-v527.js';
 import * as mestCV528 from '../../lib/mest-c-v528.js';
+import * as thwaitesV529 from '../../lib/thwaites-v529.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -951,6 +952,7 @@ const MODULES = {
   'nsofa-v526.js': nsofaV526,
   'wayne-index-v527.js': wayneIndexV527,
   'mest-c-v528.js': mestCV528,
+  'thwaites-v529.js': thwaitesV529,
   'tb-testing.js': tbTesting,
 };